requirements.txt
flask==2.3.2
instaloader==4.10.1
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
